+++ src/main/uni_gamepad.h
@@ -49,9 +49,9 @@ enum {
 // MISC_BUTTONS_ are buttons that are usually not used in the game, but are
 // helpers like "back", "home", etc.
 enum {
-  MISC_BUTTON_SYSTEM = 1 << 0,
-  MISC_BUTTON_BACK = 1 << 1,
-  MISC_BUTTON_HOME = 1 << 2,
+  MISC_BUTTON_SYSTEM = 1 << 0,  // AKA: PS, Xbox, etc.
+  MISC_BUTTON_BACK = 1 << 1,    // AKA: Select, Share, -
+  MISC_BUTTON_HOME = 1 << 2,    // AKA: Start, Options, +
 };
 
 // GAMEPAD_STATE_ are used internally to determine which button event