+++ src/main/uni_hid_parser_android.c
@@ -28,7 +28,24 @@ limitations under the License.
 
 void uni_hid_parser_android_init_report(uni_hid_device_t* d) {
   // Reset old state. Each report contains a full-state.
-  d->gamepad.updated_states = 0;
+  uni_gamepad_t* gp = &d->gamepad;
+  memset(gp, 0, sizeof(*gp));
+
+  // It is safe to set the reported states just once, here:
+  gp->updated_states = GAMEPAD_STATE_AXIS_X | GAMEPAD_STATE_AXIS_Y |
+                       GAMEPAD_STATE_AXIS_RX | GAMEPAD_STATE_AXIS_RY;
+  gp->updated_states |= GAMEPAD_STATE_BRAKE | GAMEPAD_STATE_THROTTLE;
+  gp->updated_states |= GAMEPAD_STATE_DPAD;
+  gp->updated_states |= GAMEPAD_STATE_BUTTON_X | GAMEPAD_STATE_BUTTON_Y |
+                        GAMEPAD_STATE_BUTTON_A | GAMEPAD_STATE_BUTTON_B;
+  gp->updated_states |=
+      GAMEPAD_STATE_BUTTON_TRIGGER_L | GAMEPAD_STATE_BUTTON_TRIGGER_R |
+      GAMEPAD_STATE_BUTTON_SHOULDER_L | GAMEPAD_STATE_BUTTON_SHOULDER_R;
+  gp->updated_states |=
+      GAMEPAD_STATE_BUTTON_THUMB_L | GAMEPAD_STATE_BUTTON_THUMB_R;
+  gp->updated_states |= GAMEPAD_STATE_MISC_BUTTON_BACK |
+                        GAMEPAD_STATE_MISC_BUTTON_HOME |
+                        GAMEPAD_STATE_MISC_BUTTON_SYSTEM;
 }
 
 void uni_hid_parser_android_parse_usage(uni_hid_device_t* d,
@@ -43,35 +60,33 @@ void uni_hid_parser_android_parse_usage(uni_hid_device_t* d,
       switch (usage) {
         case HID_USAGE_AXIS_X:
           gp->axis_x = uni_hid_parser_process_axis(globals, value);
-          gp->updated_states |= GAMEPAD_STATE_AXIS_X;
           break;
         case HID_USAGE_AXIS_Y:
           gp->axis_y = uni_hid_parser_process_axis(globals, value);
-          gp->updated_states |= GAMEPAD_STATE_AXIS_Y;
           break;
         case HID_USAGE_AXIS_Z:
           gp->axis_rx = uni_hid_parser_process_axis(globals, value);
-          gp->updated_states |= GAMEPAD_STATE_AXIS_RX;
           break;
         case HID_USAGE_AXIS_RZ:
           gp->axis_ry = uni_hid_parser_process_axis(globals, value);
-          gp->updated_states |= GAMEPAD_STATE_AXIS_RY;
           break;
         case HID_USAGE_HAT:
           hat = uni_hid_parser_process_hat(globals, value);
           gp->dpad = uni_hid_parser_hat_to_dpad(hat);
-          gp->updated_states |= GAMEPAD_STATE_DPAD;
           break;
         case HID_USAGE_DPAD_UP:
         case HID_USAGE_DPAD_DOWN:
         case HID_USAGE_DPAD_RIGHT:
         case HID_USAGE_DPAD_LEFT:
           uni_hid_parser_process_dpad(usage, value, &gp->dpad);
-          gp->updated_states |= GAMEPAD_STATE_DPAD;
           break;
         default:
-          logi("Android: Unsupported page: 0x%04x, usage: 0x%04x, value=0x%x\n",
-               usage_page, usage, value);
+          // Only report unsupported values if they are 1.
+          if (value)
+            logi(
+                "Android: Unsupported page: 0x%04x, usage: 0x%04x, "
+                "value=0x%x\n",
+                usage_page, usage, value);
           break;
       }
       break;
@@ -79,15 +94,17 @@ void uni_hid_parser_android_parse_usage(uni_hid_device_t* d,
       switch (usage) {
         case HID_USAGE_ACCELERATOR:
           gp->throttle = uni_hid_parser_process_pedal(globals, value);
-          gp->updated_states |= GAMEPAD_STATE_THROTTLE;
           break;
         case HID_USAGE_BRAKE:
           gp->brake = uni_hid_parser_process_pedal(globals, value);
-          gp->updated_states |= GAMEPAD_STATE_BRAKE;
           break;
         default:
-          logi("Android: Unsupported page: 0x%04x, usage: 0x%04x, value=0x%x\n",
-               usage_page, usage, value);
+          // Only report unsupported values if they are 1.
+          if (value)
+            logi(
+                "Android: Unsupported page: 0x%04x, usage: 0x%04x, "
+                "value=0x%x\n",
+                usage_page, usage, value);
           break;
       };
       break;
@@ -97,129 +114,111 @@ void uni_hid_parser_android_parse_usage(uni_hid_device_t* d,
           gp->battery = value;
           break;
         default:
-          logi("Android: Unsupported page: 0x%04x, usage: 0x%04x, value=0x%x\n",
-               usage_page, usage, value);
+          if (value)
+            logi(
+                "Android: Unsupported page: 0x%04x, usage: 0x%04x, "
+                "value=0x%x\n",
+                usage_page, usage, value);
           break;
       }
       break;
     case HID_USAGE_PAGE_BUTTON: {
       switch (usage) {
         case 0x01:  // Button A
-          if (value)
-            gp->buttons |= BUTTON_A;
-          else
-            gp->buttons &= ~BUTTON_A;
-          gp->updated_states |= GAMEPAD_STATE_BUTTON_A;
+          if (value) gp->buttons |= BUTTON_A;
           break;
         case 0x02:  // Button B
-          if (value)
-            gp->buttons |= BUTTON_B;
-          else
-            gp->buttons &= ~BUTTON_B;
-          gp->updated_states |= GAMEPAD_STATE_BUTTON_B;
+          if (value) gp->buttons |= BUTTON_B;
           break;
         case 0x03:  // non-existant button C?
           // unmapped
           break;
         case 0x04:  // Button X
-          if (value)
-            gp->buttons |= BUTTON_X;
-          else
-            gp->buttons &= ~BUTTON_X;
-          gp->updated_states |= GAMEPAD_STATE_BUTTON_X;
+          if (value) gp->buttons |= BUTTON_X;
           break;
         case 0x05:  // Button Y
-          if (value)
-            gp->buttons |= BUTTON_Y;
-          else
-            gp->buttons &= ~BUTTON_Y;
-          gp->updated_states |= GAMEPAD_STATE_BUTTON_Y;
+          if (value) gp->buttons |= BUTTON_Y;
           break;
         case 0x06:  // non-existant button Z?
           // unmapped
           break;
         case 0x07:
-          if (value)
-            gp->buttons |= BUTTON_SHOULDER_L;
-          else
-            gp->buttons &= ~BUTTON_SHOULDER_L;
-          gp->updated_states |= GAMEPAD_STATE_BUTTON_SHOULDER_L;
+          if (value) gp->buttons |= BUTTON_SHOULDER_L;
           break;
         case 0x08:
-          if (value)
-            gp->buttons |= BUTTON_SHOULDER_R;
-          else
-            gp->buttons &= ~BUTTON_SHOULDER_R;
-          gp->updated_states |= GAMEPAD_STATE_BUTTON_SHOULDER_R;
+          if (value) gp->buttons |= BUTTON_SHOULDER_R;
           break;
-        case 0x09:  // non-existant button Shoulder Z?
-        case 0x0a:  // unsupported misc button
-        case 0x0b:  // unsupported misc button
-        case 0x0c:  // unsupported misc button
+        case 0x09:
+          // Available on some Android gamepads like SteelSeries Stratus Duo.
+          if (value) gp->buttons |= BUTTON_TRIGGER_L;
+          break;
+        case 0x0a:
+          // Available on some Android gamepads like SteelSeries Stratus Duo.
+          if (value) gp->buttons |= BUTTON_TRIGGER_R;
+          break;
+        case 0x0b:
+          // Available on some Android gamepads like SteelSeries Stratus Duo.
+          if (value) gp->misc_buttons |= MISC_BUTTON_BACK;
+          break;
+        case 0x0c:
+          // Available on some Android gamepads like SteelSeries Stratus Duo.
+          if (value) gp->misc_buttons |= MISC_BUTTON_HOME;
           break;
         case 0x0d:
-          if (value)
-            gp->misc_buttons |= MISC_BUTTON_SYSTEM;
-          else
-            gp->misc_buttons &= ~MISC_BUTTON_SYSTEM;
-          gp->updated_states |= GAMEPAD_STATE_MISC_BUTTON_SYSTEM;
+          if (value) gp->misc_buttons |= MISC_BUTTON_SYSTEM;
           break;
         case 0x0e:
-          if (value)
-            gp->buttons |= BUTTON_THUMB_L;
-          else
-            gp->buttons &= ~BUTTON_THUMB_L;
-          gp->updated_states |= GAMEPAD_STATE_BUTTON_THUMB_L;
+          if (value) gp->buttons |= BUTTON_THUMB_L;
           break;
         case 0x0f:
-          if (value)
-            gp->buttons |= BUTTON_THUMB_R;
-          else
-            gp->buttons &= ~BUTTON_THUMB_R;
-          gp->updated_states |= GAMEPAD_STATE_BUTTON_THUMB_R;
+          if (value) gp->buttons |= BUTTON_THUMB_R;
           break;
         default:
-          logi("Android: Unsupported page: 0x%04x, usage: 0x%04x, value=0x%x\n",
-               usage_page, usage, value);
+          // Only report unsupported values if they are 1.
+          if (value)
+            logi(
+                "Android: Unsupported page: 0x%04x, usage: 0x%04x, "
+                "value=0x%x\n",
+                usage_page, usage, value);
           break;
       }
       break;
     }
     case HID_USAGE_PAGE_CONSUMER:
       switch (usage) {
-        case HID_USAGE_FAST_FORWARD:
-          break;
-        case HID_USAGE_REWIND:
-          break;
-        case HID_USAGE_PLAY_PAUSE:
-          break;
-        case HID_USAGE_AC_SEARCH:
-          break;
         case HID_USAGE_AC_HOME:
-          if (value)
-            gp->misc_buttons |= MISC_BUTTON_HOME;
-          else
-            gp->misc_buttons &= ~MISC_BUTTON_HOME;
-          gp->updated_states |= GAMEPAD_STATE_MISC_BUTTON_HOME;
+          // FIXME: Some devices, like SteelSeries Status Duo, use this value as
+          // BUTTON_SYSTEM. But others like Asus, use this one to report
+          // BUTTON_HOME. Instead of having a parser for Android / OUYA /
+          // 8Bitdo, we should have a HID parser and then mapping files for each
+          // VID / PID (similar to Android .kl files).
+          if (value) gp->misc_buttons |= MISC_BUTTON_HOME;
           break;
         case HID_USAGE_AC_BACK:
-          if (value)
-            gp->misc_buttons |= MISC_BUTTON_BACK;
-          else
-            gp->misc_buttons &= ~MISC_BUTTON_BACK;
-          gp->updated_states |= GAMEPAD_STATE_MISC_BUTTON_BACK;
+          if (value) gp->misc_buttons |= MISC_BUTTON_BACK;
           break;
         default:
-          logi("Android: Unsupported page: 0x%04x, usage: 0x%04x, value=0x%x\n",
-               usage_page, usage, value);
+          // Only report unsupported values if they are 1.
+          if (value)
+            logi(
+                "Android: Unsupported page: 0x%04x, usage: 0x%04x, "
+                "value=0x%x\n",
+                usage_page, usage, value);
           break;
       }
       break;
 
-    // unknown usage page
+    case 0xff01:
+      // Ignore this report. Used by some Moga devices, where page 0xff01,
+      // usage: 0x0001 is always 1.
+      if (usage == 0x01) break;
+      break;
+
     default:
-      logi("Android: Unsupported page: 0x%04x, usage: 0x%04x, value=0x%x\n",
-           usage_page, usage, value);
+      // Only report unsupported values if they are 1.
+      if (value)
+        logi("Android: Unsupported page: 0x%04x, usage: 0x%04x, value=0x%x\n",
+             usage_page, usage, value);
       break;
   }
 }